+++ src/ophyd_async/plan_stubs/__init__.py
@@ -1,3 +1,7 @@
+from .device_save_loader import (
+    load_device,
+    save_device,
+)
 from .ensure_connected import ensure_connected
 from .fly import (
     fly_and_collect,
@@ -6,8 +10,10 @@ from .fly import (
 )
 
 __all__ = [
+    "load_device",
+    "save_device",
+    "ensure_connected",
     "fly_and_collect",
     "prepare_static_seq_table_flyer_and_detectors_with_same_trigger",
     "time_resolved_fly_and_collect_with_static_seq_table",
-    "ensure_connected",
 ]
