+++ tests/plan_stubs/test_device_save_loader.py
@@ -13,6 +13,10 @@ from ophyd_async.core import (
     Device,
     SignalR,
     SignalRW,
+)
+from ophyd_async.epics.signal import epics_signal_r, epics_signal_rw
+from ophyd_async.plan_stubs.device_save_loader import (
+    all_at_once,
     get_signal_values,
     load_device,
     load_from_yaml,
@@ -21,8 +25,6 @@ from ophyd_async.core import (
     set_signal_values,
     walk_rw_signals,
 )
-from ophyd_async.core.device_save_loader import all_at_once
-from ophyd_async.epics.signal import epics_signal_r, epics_signal_rw
 
 
 class DummyChildDevice(Device):
@@ -276,9 +278,9 @@ async def test_set_signal_values_restores_value(RE: RunEngine, device, tmp_path)
     assert np.array_equal(array_value, np.array([1, 1, 1, 1, 1]))
 
 
-@patch("ophyd_async.core.device_save_loader.load_from_yaml")
-@patch("ophyd_async.core.device_save_loader.walk_rw_signals")
-@patch("ophyd_async.core.device_save_loader.set_signal_values")
+@patch("ophyd_async.plan_stubs.device_save_loader.load_from_yaml")
+@patch("ophyd_async.plan_stubs.device_save_loader.walk_rw_signals")
+@patch("ophyd_async.plan_stubs.device_save_loader.set_signal_values")
 async def test_load_device(
     mock_set_signal_values, mock_walk_rw_signals, mock_load_from_yaml, device
 ):
